Ensuring Event Handling on SIGTERM in Tritond (#19)
* Ensuring Event Handling on SIGTERM in Tritond

In response to SIGTERM, tritond will now flush any previously buffered
events.  This differs to the previous implementation as one could not be
sure when SIGTERM was handled in the event handling loop.  It was
previously possible for up to 100ms worth of events to go unpublished.

SIGINT continues to be lossy as before.

Note - This change only ensures that the existing flush logic is
applied.  More work is needed to ensure we satisfy one least once
delivery semantics.

* Code formatting fixes.