Introducing Stream Interface + Pubsub Support
Fist step towards multiplexing streams across cloud providers.  The old
Stream class has become an abstract class inherited by AWSStream
(previous Stream) and GCPStream.

Stream types are inferred by the structure of the config entry.  New
configuration syntax is additive and remain backwards compatible.
stream.get_stream serves as a factory method producing the appropriate
stream based on the config dictionary passed to it.

As of this commit, multiplexing is not yet supported.  Though a stream
can be mapped to a pubsub stream.