Added CompositeStream
New CompositeStream enables users to multiplex events to multiple
backend streams via stream identifier.  The multiplexing is lossy
in that a batch of events may be published to a subset of backends and
fail on another.  However the non-multiplexing case is also lossy, hence
we do not violate any semantics by allowing this lossy behavior.