Introducing Stream Interface + Pubsub Support (#22)
* Introducing Stream Interface + Pubsub Support

Fist step towards multiplexing streams across cloud providers.  The old
Stream class has become an abstract class inherited by AWSStream
(previous Stream) and GCPStream.

Stream types are inferred by the structure of the config entry.  New
configuration syntax is additive and remain backwards compatible.
stream.get_stream serves as a factory method producing the appropriate
stream based on the config dictionary passed to it.

As of this commit, multiplexing is not yet supported.  Though a stream
can be mapped to a pubsub stream.

* Updating README to Cover Pubsub Publication

* Formatting Fixes for README

* Added CompositeStream

New CompositeStream enables users to multiplex events to multiple
backend streams via stream identifier.  The multiplexing is lossy
in that a batch of events may be published to a subset of backends and
fail on another.  However the non-multiplexing case is also lossy, hence
we do not violate any semantics by allowing this lossy behavior.

* Renabling All Tests in Drone Pipeline

* Added Iterators for GCP and Composite Iterators

* Added README Coverage for Pubsub Consumption

* Cleaned up ack logic in GCP iterator.

* README tweaks and Removed GIT_BRANCH from Drone

* Updated README with Local Drone Testing

* Linting + Fixes from Review

* Simplified Logic in AWSStream.__init__

* Fixes From Review #2