Interrupt publishing to deprecated streams
There are only two remaining streams that are still being read from in
production (courier_event and delivery_event). We don't need to publish
any event to any other stream, so we can just throw those events away
as they are sent to tritond.

This can be done as a lightweight way to allow for the shutdown of
the kinesis streams ($$) without having to comb through thousands of
lines of code across dozens of repositories to remove all publishes to
triton. That can be done safely at a later time.